Use setters instead of build switches
